+++ app/src/main/java/com/example/tep_timeshareexchangeplatform/UI/Activity/Payment/PaymentPackage/PaymentPackageActivity.kt
@@ -320,7 +320,7 @@ class PaymentPackageActivity : BaseActivity() {
         }
     }
 
-    fun formatPrice(price: Int): String {
+    fun formatPrice(price: Long): String {
         val formatter = DecimalFormat("#,###")
         return formatter.format(price)
     }
