+++ app/src/main/java/com/example/tep_timeshareexchangeplatform/UI/Activity/Payment/PaymentRentalActivity/PaymentRentalViewModel.kt
@@ -43,7 +43,7 @@ class PaymentRentalViewModel @Inject constructor(
     private val _responseVNPAYUrl = MutableLiveData<Resource<PaymentResponse>>()
     val responseVNPAYUrl: MutableLiveData<Resource<PaymentResponse>> = _responseVNPAYUrl
     // call API to get response URL
-    fun getResponsePaymentUrl(amount: Int, orderType: String) {
+    fun getResponsePaymentUrl(amount: Long, orderType: String) {
         viewModelScope.launch {
             _responseVNPAYUrl.postValue(Resource.loading(null))
             paymentAPIRepository.getPaymentUrl(amount, orderType).let {
